split prioritization related code
