+++ tests/unit/test_core/test_mode.py
